ST-488: fixed compile time issues after vire refactor
